ml/requirements.txt
# Old requirements
# einops==0.6.1
# matplotlib==3.7.2
# numpy==1.23.5
# opencv_python==4.8.0.74
# PyYAML==6.0.1
# torch==1.11.0
# torchvision==0.12.0
# tqdm==4.65.0

certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
einops==0.6.1
fonttools==4.58.4
idna==3.10
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
matplotlib==3.7.2
natsort==8.4.0
numpy==1.23.5
opencv-python==4.8.0.74
packaging==25.0
pillow==11.2.1
pycocotools==2.0.10
pyparsing==3.0.9
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
typing_extensions==4.14.0
urllib3==2.4.0
zipp==3.23.0